[#2869] Refactor indicator.py
Create a sub-package, indicator, in akvo.rsr.models.

Split Indicator model and all its "child models" into separate modules in indicator.

There are a few functions and constants that were shared between the models. They are now in indicator.py.
(I tried to put them in indicator/__init__.py, but for some reason I couldn't get importing them to work)
