+++ akvo/rsr/models/indicator/__init__.py
@@ -1,0 +1,40 @@
+# -*- coding: utf-8 -*-
+
+# Akvo RSR is covered by the GNU Affero General Public License.
+# See more details in the license.txt file located at the root folder of the Akvo RSR module.
+# For additional details on the GNU license please see < http://www.gnu.org/licenses/agpl.html >.
+
+from disaggregation import Disaggregation
+from indicator import (Indicator, calculate_percentage, image_path, file_path, MultipleUpdateError,
+                       PERCENTAGE_MEASURE)
+from indicator_dimension import IndicatorDimension
+from indicator_label import IndicatorLabel
+from indicator_period import IndicatorPeriod
+from indicator_period_actual_dimension import IndicatorPeriodActualDimension
+from indicator_period_actual_location import IndicatorPeriodActualLocation
+from indicator_period_data import IndicatorPeriodData
+from indicator_period_data_comment import IndicatorPeriodDataComment
+from indicator_period_target_dimension import IndicatorPeriodTargetDimension
+from indicator_period_target_location import IndicatorPeriodTargetLocation
+from indicator_reference import IndicatorReference
+
+
+__all__ = [
+    'calculate_percentage',
+    'image_path',
+    'file_path',
+    'MultipleUpdateError',
+    'PERCENTAGE_MEASURE',
+    'Disaggregation',
+    'Indicator',
+    'IndicatorDimension',
+    'IndicatorLabel',
+    'IndicatorPeriod',
+    'IndicatorPeriodActualDimension',
+    'IndicatorPeriodActualLocation',
+    'IndicatorPeriodData',
+    'IndicatorPeriodDataComment',
+    'IndicatorPeriodTargetDimension',
+    'IndicatorPeriodTargetLocation',
+    'IndicatorReference',
+]
